Fixed compiler warning
git-svn-id: http://miranda.googlecode.com/svn/trunk@12867 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/srmm/msglog.c
@@ -502,7 +502,7 @@ static DWORD CALLBACK LogStreamInEvents(DWORD_PTR dwCookie, LPBYTE pbBuff, LONG 
 				*pcb = 0;
 				return 0;
 		}
-		dat->bufferLen = strlen(dat->buffer);
+		dat->bufferLen = (int)strlen(dat->buffer);
 	}
 	*pcb = min(cb, dat->bufferLen - dat->bufferOffset);
 	CopyMemory(pbBuff, dat->buffer + dat->bufferOffset, *pcb);
